requirements.txt
python-escpos
pywin32
fastapi
uvicorn